finished with delete

+++ api/urls/urls.controllers.js
@@ -4,15 +4,17 @@ const User = require("../../models/User");
 
 const baseUrl = "http:localhost:8000/urls";
 
-exports.shorten = async (req, res) => {
+exports.shorten = async (req, res, next) => {
   // create url code
+  console.log("adadadad")
   const urlCode = shortid.generate();
   try {
+    console.log(req.user)
     req.body.shortUrl = `${baseUrl}/${urlCode}`;
     req.body.urlCode = urlCode;
-    req.body.userId = req.params.userId;
+    req.body.userId = req.user._id;
     const newUrl = await Url.create(req.body);
-    await User.findByIdAndUpdate(req.params.userId, {
+    await User.findByIdAndUpdate(req.user._id, {
       $push: { urls: newUrl._id },
     });
     res.json(newUrl);
@@ -21,7 +23,7 @@ exports.shorten = async (req, res) => {
   }
 };
 
-exports.redirect = async (req, res) => {
+exports.redirect = async (req, res, next) => {
   try {
     const url = await Url.findOne({ urlCode: req.params.code });
     if (url) {
@@ -34,10 +36,14 @@ exports.redirect = async (req, res) => {
   }
 };
 
-exports.deleteUrl = async (req, res) => {
+exports.deleteUrl = async (req, res, next) => {
   try {
+
+
     const url = await Url.findOne({ urlCode: req.params.code });
+    
     if (url) {
+      if(!req.user._id.equals(url.userId)) return res.status(201).json("Unauthorized"); 
       await Url.findByIdAndDelete(url._id);
       return res.status(201).json("Deleted");
     } else {
